requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
lxml==4.6.3
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
tornado==6.1
tzlocal==3.0
urllib3==1.26.6
pdfkit==0.6.1
